move tx history operations back into unstored
